% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APA_plot.R
\name{APA_plot}
\alias{APA_plot}
\title{Visualize the 3'end APA changes between the two samples}
\usage{
APA_plot(APA_data, P_cutoff = 0.05, delta = 0.1)
}
\arguments{
\item{APA_data}{APA data generated by APA_profile function}

\item{P_cutoff}{adjusted P value cutoff to be considered as significant change}

\item{delta}{distance between two samples to be considered as significant APA events}
}
\value{
a data frame showing the significant APA change between sample1 and sample2
}
\description{
This function loads a gtf file and extract necessary information of each gene from it.
}
