% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APA_profile.R
\name{APA_profile}
\alias{APA_profile}
\title{Visualize the 3'ends of all the reads in each sample for an individual gene}
\usage{
APA_profile(
  gene_reference,
  reads,
  min = 10,
  cores = 1,
  data_type = "direct RNA"
)
}
\arguments{
\item{gene_reference}{information extracted from gtf file}

\item{reads}{information from samples}

\item{min}{minium read counts required for both samples at single gene level}

\item{cores}{number of threads used for the computation}

\item{data_type}{if the data is direct RNA}
}
\value{
a table showing the APA change for each single gene with enough depth in the data
}
\description{
This function loads a gtf file and extract necessary information of each gene from it.
}
