% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APA_profile_compare.R
\name{APA_profile_compare}
\alias{APA_profile_compare}
\title{Transcriptome wide APA analysis}
\usage{
APA_profile_compare(
  gene_reference,
  reads,
  control,
  experimental,
  min_counts = 10,
  min_reads = 5,
  min_percent = 1,
  cores = 1,
  direct_RNA = FALSE,
  internal_priming = F,
  pattern = "post",
  genome_file = NULL,
  bin = 50
)
}
\arguments{
\item{gene_reference}{information extracted from gtf file}

\item{reads}{information from samples}

\item{control}{which group in the data is used as the control group}

\item{experimental}{which group in the data is used as the experimental group}

\item{min_counts}{minimum read counts required for both groups at single gene level to perform the analysis}

\item{min_reads}{minimum read counts required for PAS calling per sample}

\item{min_percent}{minimum PAU for a PAS to be considered for downstream analysis}

\item{cores}{number of threads used for the computation}

\item{direct_RNA}{whether the data is direct RNA-Seq, TRUE or FALSE}

\item{internal_priming}{whether or not PASs subject to internal priming filtering}

\item{pattern}{to look for internal priming sequencing surrounding PAS, either "pre", "post" or "both".}

\item{genome_file}{path to a genome sequence file (.fa or .fasta)}

\item{bin}{max distance of cleavage site from PAS}
}
\value{
a table showing the APA change for each single gene with enough depth in the data
}
\description{
This function calls the PASs for each single gene, calculate PAUs for each PAS, and perform the statistics to get the P value and PolyA site usage distance between the selected two groups
}
