% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSH_plot.R
\name{CSH_plot}
\alias{CSH_plot}
\title{Visualize the 3'end APA changes between the two samples}
\usage{
CSH_plot(CSH_data, P_cutoff = 0.05, delta = 0.1)
}
\arguments{
\item{CSH_data}{CSH data generated by CSH_profile function}

\item{P_cutoff}{adjusted P value cutoff to be considered as significant change}

\item{delta}{distance between two samples to be considered as significant APA events}
}
\value{
a data frame showing the significant APA change between sample1 and sample2
}
\description{
This function loads a gtf file and extract necessary information of each gene from it.
}
