% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APA_distal_examine.R
\name{end_PAS_examine}
\alias{end_PAS_examine}
\title{Determing the 3'end APA changes between the two samples using distal or proximal PAS changes.}
\usage{
end_PAS_examine(
  PAU_data,
  reads,
  P_cutoff = 0.05,
  FC_cutoff = 1.5,
  delta = 10,
  control,
  experimental,
  position = "distal",
  type = "delta",
  cores = 1
)
}
\arguments{
\item{PAU_data}{APA data generated by APA_profile function}

\item{reads}{information from samples}

\item{P_cutoff}{adjusted P value cutoff to be considered as significant change}

\item{FC_cutoff}{fold change cutoff of PAU between two samples to be considered as significant APA events}

\item{delta}{PAU change cutoff between two samples to be considered as significant APA events}

\item{control}{which group in the data is used as the control group}

\item{experimental}{which group in the data is used as the experimental group}

\item{position}{either distal or proximal PAS is to be examined}

\item{type}{either FC or delta is used to call significant APA events}

\item{cores}{number of threads used for the computation}
}
\value{
a data table showing the significant APA change between group1 and group2
}
\description{
This function loads a gtf file and extract necessary information of each gene from it.
}
