% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_3UTR.R
\name{extract_3UTR}
\alias{extract_3UTR}
\title{Load a gtf file as reference to extract 3'end annotation}
\usage{
extract_3UTR(gtf_file, cores = 1)
}
\arguments{
\item{gtf_file}{path to the input gtf file}

\item{cores}{number of threads used for the computation}
}
\value{
a table including transcript 3'end annotation.
}
\description{
This function loads a gtf file and extract 3'end annotation of each transcript from it.
}
