% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_explore.R
\name{gene_explore}
\alias{gene_explore}
\title{Visualize the 3'ends of all the reads in each sample for an individual gene}
\usage{
gene_explore(gene_reference, reads, gene_list)
}
\arguments{
\item{gene_reference}{information extracted from gtf file}

\item{reads}{information from samples}

\item{gene_list}{name or id of a list of genes of interest}
}
\value{
plots showing the 3'end distribution of a given gene
}
\description{
Visualize the 3'ends of all the reads in each sample for an individual gene
}
