% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load a group of sample(s)}
\usage{
load_data(infile, group = group)
}
\arguments{
\item{infile}{path to the input IsoQuant output files of a group of sample(s)}

\item{group}{name or condition of sample(s) imported from infile}
}
\value{
a table including all the reads from this group of sample(s)
}
\description{
This function loads the output files from the same group of sample(s)
}
