% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gtf.R
\name{load_gtf}
\alias{load_gtf}
\title{Load a gtf file for reference}
\usage{
load_gtf(infile)
}
\arguments{
\item{infile}{path to the input gtf file}
}
\value{
a list of two lists including gene annotation and stop codon from the gtf file
}
\description{
This function loads a gtf file and extract necessary information of each gene from it.
}
