% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_matrix.R
\name{load_matrix}
\alias{load_matrix}
\title{Load  samples}
\usage{
load_matrix(infile, design = design)
}
\arguments{
\item{infile}{path to the input IsoQuant output files of group 1 samples(Control)}

\item{design}{experiment design showing the conditions of each sample file}
}
\value{
a table including all the reads from the two groups of samples
}
\description{
This function loads the IsoQuant output file including all the samples for each group and extract necessary information for the analysis.
}
