% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_samples.R
\name{load_samples}
\alias{load_samples}
\title{Load samples}
\usage{
load_samples(infile1, infile2)
}
\arguments{
\item{infile1}{path to the input IsoQuant output files of sample 1 (Control)}

\item{infile2}{path to the input IsoQuant output files of sample 2 (Treated)}
}
\value{
a S4 object of two lists including PASs information and APA changes.
}
\description{
This function loads the output files for each sample and extract necessary information for the analysis.
}
