% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAS_calling.R
\name{PAS_calling}
\alias{PAS_calling}
\title{Call PASs at single gene level}
\usage{
PAS_calling(
  gene_reference,
  reads,
  min_reads = 5,
  min_percent = 1,
  cores = 1,
  direct_RNA = FALSE,
  internal_priming = F,
  pattern = "post",
  genome_file = NULL,
  bin = 50
)
}
\arguments{
\item{gene_reference}{information extracted from gtf file}

\item{reads}{information from RNAseq samples}

\item{min_reads}{minimum read counts required at single gene level for PAS calling per sample}

\item{min_percent}{minimum percent required for a PAS to be included (0-100)}

\item{cores}{number of threads used for the computation}

\item{direct_RNA}{whether or not the data is direct RNAseq}

\item{internal_priming}{whether or not PASs subject to internal priming filtering}

\item{pattern}{to look for internal priming sequencing surrounding PAS, either "pre", "post" or "both".}

\item{genome_file}{path to a genome sequence file (.fa or .fasta)}

\item{bin}{max distance of cleavage site from PAS}
}
\value{
a table showing the top PASs for each single gene with enough depth in the data and also a bed6 file table for all the PASs
}
\description{
This function extract PolyA sites information of each gene from the provided RNAseq data.
}
