% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAU_by_sample.R
\name{PAU_by_sample}
\alias{PAU_by_sample}
\title{Call PASs at single gene level and calculate the usage of each found PAS}
\usage{
PAU_by_sample(
  gene_reference,
  reads,
  min_reads = 5,
  min_percent = 1,
  cores = 1,
  direct_RNA = F,
  internal_priming = F,
  pattern = "post",
  genome_file = NULL,
  bin = 50
)
}
\arguments{
\item{gene_reference}{information extracted from gtf file}

\item{reads}{information from RNAseq samples}

\item{min_reads}{minium reads count required at single gene level for PAS calling and PAU calculation per sample}

\item{min_percent}{minium percent required for a PAS to be included (0-100)}

\item{cores}{number of threads used for the computation}

\item{direct_RNA}{whether or not the data is direct RNAseq}

\item{internal_priming}{whether or not PASs subject to internal priming filtering}

\item{pattern}{to look for internal priming sequencing surrouding PAS, either "pre", "post" or "both".}

\item{genome_file}{path to a genome sequence file (.fa or .fasta)}

\item{bin}{max distance of cleavage site from PAS}
}
\value{
a table showing the uasage of top PASs for each single gene with enough depth in the dataset of each sample
}
\description{
This function extract PolyA sites information of each gene and calculate the usage of each found PAS from the provided RNAseq data.
}
