% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAU_test.R
\name{PAU_test}
\alias{PAU_test}
\title{Perform PAU analysis at single PAS level in each gene}
\usage{
PAU_test(PAU_data, reads, P_cutoff = 0.1)
}
\arguments{
\item{PAU_data}{output by PAU_by_sample function}

\item{reads}{information from RNAseq samples that was used to get the PAU data}

\item{P_cutoff}{cutoff of adjusted P value to be considered significant and shown as colored on the plot}
}
\value{
a table showing the usage of top PASs for each single gene with enough depth in the dataset of each sample
}
\description{
This function take PAU data output by PAU_by_sample function and compare PAU at single PAS level in each gene between two groups
}
