% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_priming_filtering.R
\name{Internal_priming}
\alias{Internal_priming}
\title{Call internal priming events}
\usage{
Internal_priming(PAS_data, pattern = "post", genome = genome)
}
\arguments{
\item{PAS_data}{date table with columns of PAS information including chrom, strand and PAS}

\item{pattern}{to look for internal priming sequencing surrounding PAS, either "pre", "post" or "both".}

\item{genome}{genome sequence loaded from a .fasta file.}
}
\value{
the same PAS_data without internal_primed "PAS".
}
\description{
This function extract PolyA sites information from the reference genome to call internal priming originate from  A-rich stretches
}
