% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_priming_filtering.R
\name{Internal_priming}
\alias{Internal_priming}
\title{Call internal priming events}
\usage{
Internal_priming(PAS_data, pattern = "pre")
}
\arguments{
\item{PAS_data}{date table with columns of PAS information including chrom, strand and PAS}

\item{pattern}{to look for internal priming sequencing surrouding PAS, either "pre", "post" or "both".}
}
\value{
the same PAS_data without internal_primed "PAS".
}
\description{
This function extract PolyA sites information from the reference genome to call internal priming originate from  A-rich stretches
}
