% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_from_bam.R
\name{load_from_bam}
\alias{load_from_bam}
\title{Load two groups of samples from bam files directly}
\usage{
load_from_bam(
  infile1,
  infile2,
  group1 = "group1",
  group2 = "group2",
  gtf_file = gtf_file,
  bambu = T,
  genome.file = F,
  cores = 1
)
}
\arguments{
\item{infile1}{path to the input IsoQuant output files of group 1 samples(Control)}

\item{infile2}{path to the input IsoQuant output files of group 2 samples(Treated)}

\item{group1}{name or condition of samples imported from infile1}

\item{group2}{name or condition of samples imported from infile2}

\item{genome.file}{genome sequence file used to align the reads in samples}

\item{cores}{number of threads used to process the data}

\item{gtf}{annotation file to assign the read to a gene in the bam file.}
}
\value{
a table including all the reads from the two groups of samples
}
\description{
This function loads the reads information from bam files for each group and extract necessary information for the analysis.
}
