% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gtf.R
\name{load_gtf}
\alias{load_gtf}
\title{Load a gtf file for reference}
\usage{
load_gtf(gtf_file, cores = 1)
}
\arguments{
\item{gtf_file}{path to the input gtf file}

\item{cores}{number of threads used for the computation}
}
\value{
a table including gene annotation, stop codon and last exon from the gtf file
}
\description{
This function loads a gtf file and extract necessary information of each gene from it.
}
