% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAS_calling.R
\name{PAS_calling}
\alias{PAS_calling}
\title{Extract top PASs used in both samples at single gene level}
\usage{
PAS_calling(
  gene_reference,
  reads,
  counts = 5,
  min = 0.05,
  cores = 1,
  data_type = "direct RNA"
)
}
\arguments{
\item{gene_reference}{information extracted from gtf file}

\item{reads}{information from samples}

\item{counts}{minium read counts required for both samples at single gene level}

\item{min}{minium fraction required for a PAS to be included}

\item{cores}{number of threads used for the computation}

\item{data_type}{if the data is direct RNA}
}
\value{
a table showing the top PASs for each single gene with enough depth in the data and also a bed file table for all the PASs
}
\description{
This function loads a gtf file and extract necessary information of each gene from it.
}
